// libs/api/app/src/util/exceptions/standard-api-errors.enum.ts
export enum StandardApiErrorsEnum {
	NOT_FOUND = 'exceptions.not-found',
	BAD_REQUEST = 'exceptions.bad-request',
	INTERNAL_SERVER_ERROR = 'exceptions.internal-server-error',
	FORBIDDEN = 'exceptions.forbidden',
	UNAUTHORIZED = 'exceptions.unauthorized',
	UNKNOWN_API_ERROR = 'exceptions.unknown',
}